Add files via upload
i added 3 languages and implanted text to speech to. the languages are french german and spanish. furthermore the text to speech is now working, able to use by clicking the play and pause button when you open the modal. the slider works you will now be also able to change the font style and sizes to your preference. 

issue1: chapter 2 translations still don't work as it shows the translation of chapter 1. 
issue2: the settingsbutton in the modal itself, the home button and the next and previous chapter button in the modal for clarification do not work. 
fix for translations: work in progress. {priority}
fix for the modal buttons: on hold {lower priority, might also remove it as it might be too many buttons that overstimulate the reader. will ask for feedback from a reader on this matter}